got code into minimal working order.
This code could use some careful refactoring. The abstration boundaries are not clear and I would say the maintainability
is low. But it has useful tests in place and it solves the use cases laid out in the kata. For a first cut, I'd say its
serviceable, as I wanted a working solution over a clean and incomplete one.